Add HearingForm for editing and creating Hearings
HearingForm is divided into multiple steps that each handle the tasks
specific to the given step. Sections step has also n amount of
substeps that are used to manage all the sections related to the
Hearing.

Idea is that HearingForm can be used to both create new Hearings and
edit existing Hearings.