Add view for creating a new Hearing and editing an existing Hearing
HearingManager is based on HearingView so that
we can reuse the existing functionality for rendering Hearing preview.

These new views provides a way to create, edit and preview Hearings.
The hearing editor has its own "draft" of the Hearing in the redux store so
that we can edit the Hearing flexibly without messing with the top level
Hearing in redux store. Editor's hearing represents the modified
state of the hearing which then may or may not be sent to backend.