Prepare for upcoming API change regarding hearing.labels
Hearing.labels are used to be sent as list of strings, but because UI
needs labels ids the API is going to change. Therefore LabelList now
supports the old response format, but also the future response format
which means labels will be list of objects containing both id and label.