Fix scrolling issue when routes are changed
Currently the page does not scroll to the top if a page gets
changed. If the user has scrolled down and clicks e.g. navigation
links, the page will stay scrolled down at the point where user
was.

This is a fix to scroll up when using routes to go to a new page.

Found the solution from here:
https://github.com/ReactTraining/react-router/blob/master/packages/react-router-dom/docs/guides/scroll-restoration.md