Provide jump links in hearing search for screen readers (#708)
There are elements between the search form and search results
that are irrelevant to search functionality. These should be
able to skip with a screen reader.

Provide a link to get to search results after the search form
and also provide a link back from the results to the form.