Update hearing comments section
The commenting section was not really accessible so this will
make it better. Buttons and other visual elements now have
descriptions for screen readers and comments are organized
as a list so it's easier to understand the hierarchy of
comments. Also the order of elements is now changed to be
more logical.

There was also some issues with the layout that should now
be fixed.

The translations were a bit confusing to the user so some
of the strings are also modified to make more sense.

Instead of showing the full date and time in every comment,
show the relative time, e.g. "2 days ago" and provide the
full date in a tooltip when hovering the element.

Comment form had labels which were not associated with the
for value of the input. This is also now fixed. This also
required to change eslint settings for labels so that it
requires either id or nesting, not both are required
at the same time now.