Update share metadata title to hearing title
Currently, the metadata title when sharing a hearing is fixed to "Kerrokantasi". This doesn't really tell the user who sees the thumbnail what the image and link is about.

I suggest displaying the title of the hearing instead. That way, when shared on facebook and twitter, users see the hearing title.