Manage focused element after clicking labels
When the user clicks a label to filter hearings
with that specific tag/label, change focus to the
select element in the hearing list search.

This is done due that with screen readers the user
loses context when they click one of the labels.

Now the focus will go on the search dropdown element
where these filters are applied.