Display small icon after external link
Turku wanted to have quick way to indicate external links with small
icons. After conversations we decided to include this to global styles.

In this case external links are defined as link that opens to new
target. This way the change was small, and no need to create that to the
editors.