Use mouse only link component for hearing images in listings
Previously there was two links in a row with the same title
that the keyboard and screen reader users could use to navigate
to a hearing or a subhearing. This is not optimal for those users
but sighted mouse users could still use the images as links to
the hearings.

Replace the actual links with mouse only clickable "links".