Fix page reload for search input onblur
Previously the onblur on the search input fired a page reload
every time when the user tabbed through the elements with a
keyboard. When page is reloaded, the focus will reset to the
page h1 element. This caused the keyboard users to remain in
an infinite loop with the h1 and search fiel when trying to
move between the elements with the tab key.

Fix the search function to not fire page reload if current
search params are the same as the inputted search or both
search params and currently inputted search are empty.