Switch map element DOM order in hearing section
The previous issue was that the content did not occure in
an intuitive order for desktop users. Since the map was present
in the DOM order before other content and it was swithed on the
right side of the content with CSS, the reading order changed
from right to left. This is bad for the usability.

The fix was to duplicate the map component for mobile and desktop
users. For desktop the map element is in the DOM after the content
that's on the left and visually it didn't change at all. For mobile
users there's the duplicate which will be visible before the main
content.