Reset form state to handle corner cases
There was a problem when after submitting form some of the data
inputted earlier were still in the components state. This caused
odd cases were picture was submitted to multiple comments. Also
when clicking the cancel button in the form just collapsed the form,
but did nothing the data.

Added state reset to few places to better handle these off cases.
This might not be the prettiest solution, but with current time
constraints this will do.