Convert end date 00:00 to 23:59
Requested feature so that picking an end date defaults to end of day
(23:59:59) instead of start of day (00:00:00). react-datetime didn't
seem to provide any good way of doing this. Therefore we always convert
an end datetime that isSame as start of day into end of day. An
unfortunate side effect is that it is no longer possible to select
"00:00" as the end time, the earliest possible time for the day is now
"00:01".