fix: fix save hearing as copy
- cleanHearing -> prepareHearingForSave
- Remove IDs from *all* objects, not just
  the section
- Add new reference_id field
- Mark freshly uploaded files with isNew

reference_id lets you use an existing
instance as a base for a new instance.

Newly uploaded files are marked with isNew
to differentiate between existing instances
and freshly uploaded files.
E.g. if you edit an existing hearing with
a file, add a new file, and then save it
as a copy, then you want to copy the
existing file and keep the new file's ID
(since it's doesn't belong to any hearing
yet)